requirements.txt
pandas
gitpython
pytest
oxenai